r8842/sofa-dev : BUGFIX: MechanicalSetConstraint visitor was not calling buildConstraintMatrix on constraints below mappings with isMechanical="false". Also add index checks in FullVector.
 //partial sync (4/8 changes)//

+++ framework/sofa/helper/vector.cpp
@@ -40,7 +40,7 @@ DEBUG_OUT_V(int cptid = 0);
 
 void SOFA_HELPER_API vector_access_failure(const void* vec, unsigned size, unsigned i, const std::type_info& type)
 {
-    std::cerr << "ERROR in vector<"<<gettypename(type)<<"> " << std::hex << (long)vec << std::dec << " size " << size << " : invalid index " << i << std::endl;
+    std::cerr << "ERROR in vector<"<<gettypename(type)<<"> " << std::hex << (long)vec << std::dec << " size " << size << " : invalid index " << (int)i << std::endl;
     BackTrace::dump();
     assert(i < size);
 }
